user_auth/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-allauth==0.63.2
sqlparse==0.5.0
typing_extensions==4.12.0
